Move `WolfSession` into `session.rs`
